kubevirt: rename vm status file
