+++ frontend/packages/kubevirt-plugin/src/statuses/vm/vm-status.ts
